refactor(react-native): change CallParticipantsView to CallParticipantsGridView and conditionally render the grid and spotlight component
